feat: Add GitHub Actions workflow for generating changelog
A new GitHub Actions workflow has been added to the repository. This workflow generates a changelog based on the commits in the master branch and attaches it to the release. The changelog is generated using the git-cliff tool with a custom configuration file. This will help provide a comprehensive overview of the changes made in each release.