feat: Update githubHandler to include release version in text part of the slack message
The githubHandler file has been modified to include the release version in the Slack message when a new tag is pushed. This ensures that the Slack message provides information about the released version. If the ref starts with "refs/tags/", the message will be "released new version: [version]". Otherwise, it will be "committed: [commit message]".

Co-authored-by: [Author Name]